Re-organized generated_parsers. Added a keep-erros option to conversion function
