+++ src/compat_impl/compat_inspec_1_0.ts
@@ -2,9 +2,9 @@
  * A lot of information/behaviour is shared between the profile and result version so we use a single abstract superclass
  */
 
-import { ExecJSONControl as ResultControl_1_0 } from "../generated_parsers/exec-json";
-import { ProfileJSONControl as ProfileControl_1_0 } from "../generated_parsers/profile-json";
-import { ControlResult as ControlResult_1_0 } from "../generated_parsers/exec-json";
+import { ExecJSONControl as ResultControl_1_0 } from "../generated_parsers/v_1_0/exec-json";
+import { ProfileJSONControl as ProfileControl_1_0 } from "../generated_parsers/v_1_0/profile-json";
+import { ControlResult as ControlResult_1_0 } from "../generated_parsers/v_1_0/exec-json";
 
 import {
   HDFControl,
