+++ src/fileparse.ts
@@ -1,8 +1,8 @@
 // Import all of our parsers
 // V 1.0
-import * as EXEC_JSON_1_0 from "./generated_parsers/exec-json";
-import * as EXEC_JSON_MIN_1_0 from "./generated_parsers/exec-jsonmin";
-import * as PROFILE_JSON_1_0 from "./generated_parsers/profile-json";
+import * as EXEC_JSON_1_0 from "./generated_parsers/v_1_0/exec-json";
+import * as EXEC_JSON_MIN_1_0 from "./generated_parsers/v_1_0/exec-jsonmin";
+import * as PROFILE_JSON_1_0 from "./generated_parsers/v_1_0/profile-json";
 
 // TODO: Any future versions
 
@@ -13,6 +13,7 @@ export interface ConversionResult {
   "1_0_ExecJson"?: EXEC_JSON_1_0.ExecJSON;
   "1_0_ExecJsonMin"?: EXEC_JSON_MIN_1_0.ExecJsonmin;
   "1_0_ProfileJson"?: PROFILE_JSON_1_0.ProfileJSON;
+  errors?: any[];
 }
 
 /**
@@ -20,7 +21,10 @@ export interface ConversionResult {
  *
  * @param json_text The json file contents
  */
-export function convertFile(json_text: string): ConversionResult {
+export function convertFile(
+  json_text: string,
+  keep_errors = false
+): ConversionResult {
   // Establish result
   const result: ConversionResult = {};
 
@@ -52,6 +56,12 @@ export function convertFile(json_text: string): ConversionResult {
   }
 
   // errors.forEach(e => console.warn(e));
+  if (keep_errors) {
+    return {
+      ...result,
+      errors
+    };
+  }
   throw new Error(
     "Unable to convert input json. " +
       "This usually means the file is malformed - please check that the software that generated it is up to date, " +
